+++ bravo/plugins/commands/debug.py
@@ -4,7 +4,29 @@ from bravo.ibravo import IConsoleCommand, IChatCommand
 
 from bravo.parameters import factory
 
+# Trivial hello-world command.
+# If this is ever modified, please also update the documentation;
+# docs/extending.rst includes this verbatim in order to demonstrate authoring
+# commands.
+class Hello(object):
+    """
+    Say hello to the world.
+    """
+
+    implements(IChatCommand)
+
+    def chat_command(self, username, parameters):
+        greeting = "Hello, %s!" % username
+        yield greeting
+
+    name = "hello"
+    aliases = tuple()
+    usage = ""
+
 class Meliae(object):
+    """
+    Dump a Meliae snapshot to disk.
+    """
 
     implements(IConsoleCommand)
 
@@ -23,9 +45,11 @@ class Meliae(object):
     name = "dump-memory"
     aliases = tuple()
     usage = "<filename>"
-    info = "Dump a JSON snapshot of memory usage using Meliae"
 
 class Status(object):
+    """
+    Print a short summary of the world's status.
+    """
 
     implements(IConsoleCommand)
 
@@ -46,9 +70,11 @@ class Status(object):
     name = "status"
     aliases = tuple()
     usage = ""
-    info = "Print a quick summary of the server's status"
 
 class Colors(object):
+    """
+    Paint with all the colors of the wind.
+    """
 
     implements(IChatCommand)
 
@@ -62,9 +88,11 @@ class Colors(object):
     name = "colors"
     aliases = tuple()
     usage = ""
-    info = "Print the colors"
 
 class Rain(object):
+    """
+    Perform a rain dance.
+    """
 
     implements(IChatCommand)
 
@@ -82,8 +110,8 @@ class Rain(object):
     name = "rain"
     aliases = tuple()
     usage = "<state>"
-    info = "Do a rain dance"
 
+hello = Hello()
 meliae = Meliae()
 status = Status()
 colors = Colors()