+++ CHANGELOG.rst
@@ -4,7 +4,18 @@ next
 Compatibility
 -------------
 
+Configuration
+^^^^^^^^^^^^^
+
 * Configuration for factories now uses endpoints
+
+Plugins
+^^^^^^^
+
+* All plugins no longer need to implement ``twisted.plugin.IPlugin``
+* Plugins may live in subpackages of ``bravo.plugins``
+* Command plugins are documented in docstrings instead of the ``info``
+  attribute
 
 Features
 --------